simple mpi support
